some optimization
git-svn-id: https://kforge.ros.org/fcl/fcl_ros@138 253336fb-580f-4252-a368-f3cef5a2a82b

+++ trunk/fcl/include/fcl/collision_data.h
@@ -59,15 +59,15 @@ struct CostSource
 struct CollisionRequest
 {
   bool exhaustive;
-  int num_max_contacts;
+  size_t num_max_contacts;
   bool enable_contact;
-  int num_max_cost_sources;
+  size_t num_max_cost_sources;
   bool enable_cost;
 
   CollisionRequest(bool exhaustive_ = false,
-                   unsigned int num_max_contacts_ = 1,
+                   size_t num_max_contacts_ = 1,
                    bool enable_contact_ = false,
-                   unsigned int num_max_cost_sources_ = 1,
+                   size_t num_max_cost_sources_ = 1,
                    bool enable_cost_ = false) : exhaustive(exhaustive_),
                                                 num_max_contacts(num_max_contacts_),
                                                 enable_contact(enable_contact_),
